pyproject.toml
[project]
name = "databricks-async-client"
version = "0.1.0"
description = ""

readme = "README.md"
#requires-python = ">=3.14,<4.0.0"
requires-python = "==3.12.*"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "fastapi[standard] (>=0.121.0,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "redis (>=7.0.1,<8.0.0)"
]

[tool.poetry]
packages = [{include = "rest_api", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "jupyter (>=1.1.1,<2.0.0)",
    #"pyspark (>=3.5.0,<4.0.0)",  # For Spark Connect local dev
    "databricks-vectorsearch==0.61",
    #"grpcio (>=1.48.1,<2.0.0)", # Required for Spark Connect
    #"grpcio-status (>=1.48.1,<2.0.0)",  # Required for Spark Connect
    "databricks-connect==16.4.*"
]
